services/model-deployer/pyproject.toml
[tool.poetry]
name = "model-deployer"
version = "0.1.0"
description = ""
authors = ["Liam Monninger"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
openai = "^0.27.2"
fastapi = "^0.95.0"
python-dotenv = "^1.0.0"
uvicorn = "^0.21.1"
nest-asyncio = "^1.5.6"
websockets = "^11.0"
google-cloud-run = "^0.7.1"
redis = "^4.5.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
